Course now has a container and is presentational
