+++ src/components/Dashboard/Course/New/Info.jsx
